Refactoring the gui code
