+++ Editor/LocalisationConfigGenerator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d851144310b1439a9e1bb61f1876b08d
+timeCreated: 1562245888
