+++ Editor/LocalisationTable.Editor.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using DUCK.Localisation.Editor.Window;
 using UnityEditor;
 using UnityEngine;
