+++ Editor/Window/LocalisationWindow.CodeGenerationGUI.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fe57f1f349fb4f32b93fc5b084d3ca8d
+timeCreated: 1562246644
