+++ Editor/Window/LocalisationWindow.SettingsGUI.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b6a2c30721cf4d03a5adf0908a541d8e
+timeCreated: 1562246185
