+++ Editor/Window/LocalisationWindow.SettingsGui.cs
@@ -1,0 +1,43 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace DUCK.Localisation.Editor.Window
+{
+	public partial class LocalisationWindow
+	{
+		public class SettingsGui
+		{
+			public void Draw()
+			{
+				// Draw title
+				EditorGUILayout.Space();
+				GUILayout.Label("Settings", EditorStyles.boldLabel);
+
+				// Draw settings field
+				var settings = LocalisationSettings.Current;
+
+				settings =
+					(LocalisationSettings) EditorGUILayout.ObjectField(
+						"Settings",
+						settings,
+						typeof(LocalisationSettings),
+						false);
+
+				LocalisationSettings.Current = settings;
+
+				// if no settings then draw the help box and create button
+				if (settings == null)
+				{
+					EditorGUILayout.HelpBox(
+						"Please specify a settings object, or alternatively create a new one with the button below",
+						MessageType.Info);
+
+					if (GUILayout.Button("Create new LocalisationSettings"))
+					{
+						LocalisationSettings.Create();
+					}
+				}
+			}
+		}
+	}
+}
