+++ Editor/Window/LocalisationWindow.TablesGui.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d8927e293fe84e3d9f29f44cb12a20b8
+timeCreated: 1562246812
